requirements.txt
Django>=4.2 
Pillow>=10.0.0
# Django==4.1 
# Pillow==9.5.0
psycopg2-binary>=2.9
gunicorn>=20.1
python-dotenv>=1.0.0
dj-database-url>=2.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
httpx>=0.24.0
apscheduler>=3.10.0
whitenoise>=6.0